refactor: lexwebapp DX improvements — path aliases, error handling, module splits
Phase 1: Add @/* path aliases (tsconfig + vite), ErrorBoundary with Ukrainian UI
Phase 2: Split 531-line api-client.ts into domain modules under src/utils/api/
Phase 3: Replace 20 `catch(error: any)` with typed `catch(error: unknown)` + getErrorMessage util
Phase 4: Extract useDocumentData + useDocumentActions hooks from 1488-line DocumentsPage
Phase 5: Extract MCPService formatters and response transformers into src/services/api/mcp/
