+++ lexwebapp/src/utils/api-client.ts
@@ -1,531 +1,6 @@
 /**
- * API Client Utility
- * Axios instance with authentication and error handling
+ * Re-export shim — preserves all existing imports.
+ * New code should import from '@/utils/api' or '@/utils/api/client' directly.
  */
-
-import axios, { AxiosError, AxiosInstance, InternalAxiosRequestConfig } from 'axios';
-import { showToast } from './toast';
-
-// Validate API URL against allowed origins
-const ALLOWED_API_ORIGINS = [
-  'https://legal.org.ua',
-  'https://stage.legal.org.ua',
-  'http://localhost:3000',
-  'http://localhost:8080',
-];
-
-const API_URL = import.meta.env.VITE_API_URL || 'http://localhost:3000';
-
-if (!ALLOWED_API_ORIGINS.some(origin => API_URL.startsWith(origin))) {
-  throw new Error(
-    `Invalid VITE_API_URL: "${API_URL}". Must start with one of: ${ALLOWED_API_ORIGINS.join(', ')}`
-  );
-}
-
-// Token refresh state
-let isRefreshing = false;
-let failedQueue: Array<{
-  resolve: (token: string) => void;
-  reject: (error: any) => void;
-}> = [];
-
-function processQueue(error: any, token: string | null = null) {
-  failedQueue.forEach(({ resolve, reject }) => {
-    if (error) {
-      reject(error);
-    } else {
-      resolve(token!);
-    }
-  });
-  failedQueue = [];
-}
-
-function forceLogout() {
-  localStorage.removeItem('auth_token');
-  localStorage.removeItem('user');
-  window.location.href = '/login';
-}
-
-// Create axios instance
-const apiClient: AxiosInstance = axios.create({
-  baseURL: API_URL,
-  headers: {
-    'Content-Type': 'application/json',
-  },
-  timeout: 30000, // 30 seconds
-});
-
-// Request interceptor - attach JWT token
-apiClient.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem('auth_token');
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-    return config;
-  },
-  (error) => {
-    return Promise.reject(error);
-  }
-);
-
-// Response interceptor - handle errors with token refresh
-apiClient.interceptors.response.use(
-  (response) => response,
-  async (error: AxiosError<any>) => {
-    // Network error
-    if (!error.response) {
-      showToast.error('Network error. Please check your connection.');
-      return Promise.reject(error);
-    }
-
-    const originalRequest = error.config as InternalAxiosRequestConfig & { _retry?: boolean };
-    const { status, data } = error.response;
-
-    // 401 Unauthorized - attempt token refresh before logging out
-    if (status === 401) {
-      // Skip refresh for auth endpoints to avoid infinite loops
-      const url = originalRequest?.url || '';
-      if (url.includes('/auth/refresh') || url.includes('/auth/login')) {
-        forceLogout();
-        return Promise.reject(error);
-      }
-
-      // Already retried this request — give up
-      if (originalRequest._retry) {
-        forceLogout();
-        return Promise.reject(error);
-      }
-
-      // If a refresh is already in progress, queue this request
-      if (isRefreshing) {
-        return new Promise<string>((resolve, reject) => {
-          failedQueue.push({ resolve, reject });
-        }).then((token) => {
-          originalRequest.headers.Authorization = `Bearer ${token}`;
-          return apiClient(originalRequest);
-        }).catch((err) => {
-          return Promise.reject(err);
-        });
-      }
-
-      originalRequest._retry = true;
-      isRefreshing = true;
-
-      try {
-        const { data: refreshData } = await axios.post(
-          `${API_URL}/auth/refresh`,
-          {},
-          {
-            headers: {
-              Authorization: `Bearer ${localStorage.getItem('auth_token')}`,
-              'Content-Type': 'application/json',
-            },
-            timeout: 10000,
-          }
-        );
-
-        const newToken = refreshData.token;
-        localStorage.setItem('auth_token', newToken);
-
-        // Retry all queued requests with the new token
-        processQueue(null, newToken);
-
-        // Retry the original request
-        originalRequest.headers.Authorization = `Bearer ${newToken}`;
-        return apiClient(originalRequest);
-      } catch (refreshError) {
-        processQueue(refreshError, null);
-        showToast.error('Session expired. Please login again.');
-        forceLogout();
-        return Promise.reject(refreshError);
-      } finally {
-        isRefreshing = false;
-      }
-    }
-
-    // 402 Payment Required - insufficient balance
-    if (status === 402) {
-      const message = data?.message || 'Insufficient balance. Please top up your account.';
-      showToast.error(message, 5000);
-
-      return Promise.reject(error);
-    }
-
-    // 429 Too Many Requests - rate limit exceeded
-    if (status === 429) {
-      // Skip toast for upload endpoints — UploadManager handles 429 retry internally
-      const url = error.config?.url || '';
-      if (!url.includes('/upload/')) {
-        const message = data?.message || 'Rate limit exceeded. Please try again later.';
-        showToast.error(message);
-      }
-      return Promise.reject(error);
-    }
-
-    // 500+ Server errors
-    if (status >= 500) {
-      showToast.error('Server error. Please try again later.');
-      return Promise.reject(error);
-    }
-
-    // Other errors - show message from server
-    if (data?.message) {
-      showToast.error(data.message);
-    } else {
-      showToast.error('An error occurred. Please try again.');
-    }
-
-    return Promise.reject(error);
-  }
-);
-
-export default apiClient;
-
-/**
- * API Service Object
- * Centralized API methods for auth, billing, payments, team, and tools
- */
-export const api = {
-  // Auth
-  auth: {
-    getMe: () => apiClient.get('/auth/me'),
-    logout: () => apiClient.post('/auth/logout'),
-    refreshToken: () => apiClient.post('/auth/refresh'),
-    updateProfile: (data: { name?: string; picture?: string }) =>
-      apiClient.put('/auth/profile', data),
-  },
-
-  // Billing
-  billing: {
-    getBalance: () => apiClient.get('/api/billing/balance'),
-    getSettings: () => apiClient.get('/api/billing/settings'),
-    getHistory: (params?: { limit?: number; offset?: number; type?: string }) =>
-      apiClient.get('/api/billing/history', { params }),
-    updateSettings: (data: {
-      daily_limit_usd?: number;
-      monthly_limit_usd?: number;
-      email_notifications?: boolean;
-      notify_low_balance?: boolean;
-      notify_payment_success?: boolean;
-      notify_payment_failure?: boolean;
-      notify_monthly_report?: boolean;
-      low_balance_threshold_usd?: number;
-    }) => apiClient.put('/api/billing/settings', data),
-    getEmailPreferences: () => apiClient.get('/api/billing/email-preferences'),
-    getInvoices: (params?: { limit?: number; offset?: number }) =>
-      apiClient.get('/api/billing/invoices', { params }),
-    downloadInvoicePDF: (invoiceNumber: string) =>
-      apiClient.get(`/api/billing/invoices/${invoiceNumber}/pdf`, { responseType: 'blob' }),
-    testEmail: () => apiClient.post('/api/billing/test-email'),
-    getStatistics: (period: string = '30d') =>
-      apiClient.get(`/api/billing/statistics?period=${period}`),
-    getUsageChart: (days: number = 30) =>
-      apiClient.get(`/api/billing/usage-chart?days=${days}`),
-    getPaymentMethods: () => apiClient.get('/api/billing/payment-methods'),
-    addPaymentMethod: (data: any) =>
-      apiClient.post('/api/billing/payment-methods', data),
-    removePaymentMethod: (id: string) =>
-      apiClient.delete(`/api/billing/payment-methods/${id}`),
-    setPrimaryPaymentMethod: (id: string) =>
-      apiClient.put(`/api/billing/payment-methods/${id}/primary`, {}),
-    upgradePlan: (planId: string) =>
-      apiClient.put('/api/billing/settings', { pricingTier: planId }),
-    getPricingInfo: () => apiClient.get('/api/billing/pricing-info'),
-  },
-
-  // Payments
-  payment: {
-    createMonobank: (data: { amount_uah: number; redirect_url?: string }) =>
-      apiClient.post('/api/billing/payment/monobank/create', data),
-    createMetaMask: (data: { amount_usd: number; network: string; token: string }) =>
-      apiClient.post('/api/billing/payment/metamask/create', data),
-    verifyMetaMask: (data: { paymentIntentId: string; txHash: string }) =>
-      apiClient.post('/api/billing/payment/metamask/verify', data),
-    createBinancePay: (data: { amount_usd: number }) =>
-      apiClient.post('/api/billing/payment/binance-pay/create', data),
-    getAvailableProviders: () =>
-      apiClient.get('/api/billing/payment/available-providers'),
-    getStatus: (provider: string, paymentId: string) =>
-      apiClient.get(`/api/billing/payment/${provider}/${paymentId}/status`),
-  },
-
-  // Team Management
-  team: {
-    getMembers: () => apiClient.get('/api/team/members'),
-    getStats: () => apiClient.get('/api/team/stats'),
-    getOrganization: () => apiClient.get('/api/team/organization'),
-    createOrganization: (data: { name: string; taxId?: string; contactEmail?: string; description?: string }) =>
-      apiClient.post('/api/team/organization', data),
-    inviteMember: (email: string, role: string) =>
-      apiClient.post('/api/team/invite', { email, role }),
-    updateMember: (memberId: string, data: any) =>
-      apiClient.put(`/api/team/members/${memberId}`, data),
-    removeMember: (memberId: string) =>
-      apiClient.delete(`/api/team/members/${memberId}`),
-    resendInvite: (memberId: string) =>
-      apiClient.post(`/api/team/members/${memberId}/resend-invite`, {}),
-  },
-
-  // Tools
-  tools: {
-    execute: (toolName: string, params: any) =>
-      apiClient.post(`/api/tools/${toolName}`, params),
-  },
-
-  // Conversations
-  conversations: {
-    create: (title?: string) => apiClient.post('/api/conversations', { title }),
-    list: (params?: { limit?: number; offset?: number }) =>
-      apiClient.get('/api/conversations', { params }),
-    get: (id: string) => apiClient.get(`/api/conversations/${id}`),
-    rename: (id: string, title: string) =>
-      apiClient.put(`/api/conversations/${id}`, { title }),
-    delete: (id: string) => apiClient.delete(`/api/conversations/${id}`),
-    addMessage: (conversationId: string, message: {
-      role: 'user' | 'assistant';
-      content: string;
-      thinking_steps?: any[];
-      decisions?: any[];
-      citations?: any[];
-      documents?: any[];
-      cost_summary?: any;
-    }) => apiClient.post(`/api/conversations/${conversationId}/messages`, message),
-  },
-
-  // Documents
-  documents: {
-    getFolders: (prefix?: string) =>
-      apiClient.get('/api/documents/folders', { params: { prefix } }),
-    deleteFolder: (folderPath: string) =>
-      apiClient.delete(`/api/documents/folders/${encodeURIComponent(folderPath)}`),
-    getById: (id: string) => apiClient.get(`/api/documents/${id}`),
-    delete: (id: string) => apiClient.delete(`/api/documents/${id}`),
-    restore: (id: string) => apiClient.patch(`/api/documents/${id}/restore`),
-    update: (id: string, data: { full_text?: string; title?: string; type?: string }) =>
-      apiClient.patch(`/api/documents/${id}`, data),
-    move: (id: string, folderPath: string) =>
-      apiClient.post(`/api/documents/${id}/move`, { folderPath }),
-    getStats: () => apiClient.get('/api/documents/stats'),
-    getPreviewUrl: (id: string) => apiClient.get(`/api/documents/${id}/preview`),
-    startClassification: (params: { concurrency?: number; documentIds?: string[] }) =>
-      apiClient.post('/api/documents/classify', params),
-    getClassificationJob: (jobId: string) =>
-      apiClient.get(`/api/documents/classify/${jobId}`),
-    cancelClassification: (jobId: string) =>
-      apiClient.post(`/api/documents/classify/${jobId}/cancel`),
-    dismissClassification: () =>
-      apiClient.post('/api/documents/classify/dismiss'),
-  },
-
-  // Admin
-  admin: {
-    getDataSources: (section?: string) =>
-      apiClient.get('/api/admin/data-sources', { params: section ? { section } : undefined }),
-    getOverview: () => apiClient.get('/api/admin/stats/overview'),
-    getRevenueChart: (days: number = 30) =>
-      apiClient.get(`/api/admin/stats/revenue-chart?days=${days}`),
-    getTierDistribution: () => apiClient.get('/api/admin/stats/tier-distribution'),
-    getUsers: (params?: { limit?: number; offset?: number; search?: string; tier?: string; status?: string }) =>
-      apiClient.get('/api/admin/users', { params }),
-    getUser: (id: string) => apiClient.get(`/api/admin/users/${id}`),
-    updateUserTier: (id: string, tier: string) =>
-      apiClient.put(`/api/admin/users/${id}/tier`, { tier }),
-    adjustBalance: (id: string, amount: number, reason: string) =>
-      apiClient.post(`/api/admin/users/${id}/adjust-balance`, { amount, reason }),
-    updateLimits: (id: string, limits: { dailyLimitUsd?: number; monthlyLimitUsd?: number }) =>
-      apiClient.put(`/api/admin/users/${id}/limits`, limits),
-    getTransactions: (params?: { limit?: number; offset?: number; type?: string; status?: string; userId?: string }) =>
-      apiClient.get('/api/admin/transactions', { params }),
-    refundTransaction: (id: string, reason: string) =>
-      apiClient.post(`/api/admin/transactions/${id}/refund`, { reason }),
-    getUsageAnalytics: (days: number = 30) =>
-      apiClient.get(`/api/admin/analytics/usage?days=${days}`),
-    getCohorts: () => apiClient.get('/api/admin/analytics/cohorts'),
-    getCostBreakdown: (days: number = 30) =>
-      apiClient.get('/api/admin/stats/cost-breakdown', { params: { days } }),
-    getUsersCostsSummary: (days: number = 30) =>
-      apiClient.get('/api/admin/stats/users-costs', { params: { days } }),
-    getUserActivity: (params?: { hours?: number; limit?: number; recent?: number }) =>
-      apiClient.get('/api/admin/users/activity', { params }),
-    getUserRequests: (userId: string, params?: { limit?: number; offset?: number; days?: number }) =>
-      apiClient.get(`/api/admin/users/${userId}/requests`, { params }),
-    getUploadMetrics: () => apiClient.get('/api/admin/upload-metrics'),
-    getRecentCourtDocs: (days: number = 30, limit: number = 5) =>
-      apiClient.get('/api/admin/court-documents/recent', { params: { days, limit } }),
-    runDocumentCompletenessCheck: () =>
-      apiClient.post('/api/admin/document-completeness-check'),
-    startBackfillFulltext: (params?: { justice_kind_code?: string; limit?: number; concurrency?: number; proxy?: string }) =>
-      apiClient.post('/api/admin/backfill-fulltext', params),
-    getBackfillStatus: (jobId?: string) =>
-      jobId
-        ? apiClient.get(`/api/admin/backfill-fulltext/${jobId}`)
-        : apiClient.get('/api/admin/backfill-fulltext'),
-    stopBackfill: (jobId: string) =>
-      apiClient.post(`/api/admin/backfill-fulltext/${jobId}/stop`),
-    deleteBackfillJob: (jobId: string) =>
-      apiClient.delete(`/api/admin/backfill-fulltext/${jobId}`),
-    getTrafficMetrics: (range: string = '1h') =>
-      apiClient.get('/api/admin/metrics/traffic', { params: { range } }),
-    getLatencyMetrics: (range: string = '1h') =>
-      apiClient.get('/api/admin/metrics/latency', { params: { range } }),
-    getServicesHealth: () =>
-      apiClient.get('/api/admin/metrics/services'),
-    getSystemMetrics: () =>
-      apiClient.get('/api/admin/metrics/system'),
-
-    // Billing management
-    getBillingTiers: () =>
-      apiClient.get('/api/admin/billing/tiers'),
-    updateBillingTier: (idOrKey: string, data: any) =>
-      apiClient.put(`/api/admin/billing/tiers/${idOrKey}`, data),
-    setDefaultTier: (id: string) =>
-      apiClient.put(`/api/admin/billing/tiers/${id}/default`),
-    deleteBillingTier: (id: string) =>
-      apiClient.delete(`/api/admin/billing/tiers/${id}`),
-    getVolumeDiscounts: () =>
-      apiClient.get('/api/admin/billing/volume-discounts'),
-    updateVolumeDiscounts: (thresholds: any[]) =>
-      apiClient.put('/api/admin/billing/volume-discounts', { thresholds }),
-    getOrganizations: () =>
-      apiClient.get('/api/admin/billing/organizations'),
-    getOrganization: (id: string) =>
-      apiClient.get(`/api/admin/billing/organizations/${id}`),
-    updateOrganization: (id: string, data: any) =>
-      apiClient.put(`/api/admin/billing/organizations/${id}`, data),
-    getSubscriptions: (params?: { limit?: number; offset?: number; status?: string; tier?: string }) =>
-      apiClient.get('/api/admin/billing/subscriptions', { params }),
-    createSubscription: (data: any) =>
-      apiClient.post('/api/admin/billing/subscriptions', data),
-    updateSubscription: (id: string, data: any) =>
-      apiClient.put(`/api/admin/billing/subscriptions/${id}`, data),
-    deleteSubscription: (id: string) =>
-      apiClient.delete(`/api/admin/billing/subscriptions/${id}`),
-    cancelSubscription: (id: string, reason: string) =>
-      apiClient.put(`/api/admin/billing/subscriptions/${id}/cancel`, { reason }),
-    activateSubscription: (id: string) =>
-      apiClient.put(`/api/admin/billing/subscriptions/${id}/activate`),
-    getSubscriptionStats: () =>
-      apiClient.get('/api/admin/billing/subscription-stats'),
-
-    // Container metrics (cAdvisor)
-    getContainerMetrics: (range: string = '1h') =>
-      apiClient.get('/api/admin/metrics/containers', { params: { range } }),
-
-    // Infrastructure dashboards
-    getInfrastructureMetrics: (range: string = '1h') =>
-      apiClient.get('/api/admin/metrics/infrastructure', { params: { range } }),
-    getUploadPipelineMetrics: (range: string = '1h') =>
-      apiClient.get('/api/admin/metrics/upload-pipeline', { params: { range } }),
-    getBackendDetailMetrics: (range: string = '1h') =>
-      apiClient.get('/api/admin/metrics/backend-detail', { params: { range } }),
-    getCostRealtimeMetrics: (range: string = '6h') =>
-      apiClient.get('/api/admin/metrics/cost-realtime', { params: { range } }),
-    getUserTags: (userId: string) =>
-      apiClient.get(`/api/admin/users/${userId}/tags`),
-    toggleCryptoTag: (userId: string, enable: boolean) =>
-      enable
-        ? apiClient.put(`/api/admin/users/${userId}/tags/crypto`)
-        : apiClient.delete(`/api/admin/users/${userId}/tags/crypto`),
-    toggleTestTag: (userId: string, enable: boolean) =>
-      enable
-        ? apiClient.put(`/api/admin/users/${userId}/tags/test`)
-        : apiClient.delete(`/api/admin/users/${userId}/tags/test`),
-    createTestUser: (data: { email: string; name?: string; password: string; credits: number }) =>
-      apiClient.post('/api/admin/test-users', data),
-    resetUserPassword: (id: string) =>
-      apiClient.post(`/api/admin/users/${id}/reset-password`),
-
-    getImportSamples: (hours: number = 24, limit: number = 5) =>
-      apiClient.get('/api/admin/import-samples', { params: { hours, limit } }),
-
-    // System configuration
-    getConfig: () => apiClient.get('/api/admin/config'),
-    updateConfig: (key: string, value: string) =>
-      apiClient.put(`/api/admin/config/${key}`, { value }),
-    resetConfig: (key: string) =>
-      apiClient.delete(`/api/admin/config/${key}`),
-    getDBCompare: () =>
-      apiClient.get('/api/admin/db-compare'),
-
-    // Court registry scraper
-    startCourtScraper: (params?: {
-      justice_kind?: string;
-      justice_kind_id?: string;
-      doc_form?: string;
-      date_from?: string;
-      max_docs?: number;
-      concurrency?: number;
-      proxy?: string;
-    }) => apiClient.post('/api/admin/scrape-court-registry', params),
-    getCourtScraperStatus: (jobId?: string) =>
-      jobId
-        ? apiClient.get(`/api/admin/scrape-court-registry/${jobId}`)
-        : apiClient.get('/api/admin/scrape-court-registry'),
-    stopCourtScraper: (jobId: string) =>
-      apiClient.post(`/api/admin/scrape-court-registry/${jobId}/stop`),
-    deleteCourtScraperJob: (jobId: string) =>
-      apiClient.delete(`/api/admin/scrape-court-registry/${jobId}`),
-    getAllScraperJobs: () =>
-      apiClient.get('/api/admin/scrape-court-registry/all'),
-    getRegistryCoverageMap: (years?: number) =>
-      apiClient.get(`/api/admin/registry-coverage-map${years ? `?years=${years}` : ''}`),
-
-    // Service pricing
-    getServicePricing: () =>
-      apiClient.get('/api/admin/service-pricing'),
-    updateServicePricing: (id: string, data: { price_usd: number; notes?: string; is_active?: boolean }) =>
-      apiClient.put(`/api/admin/service-pricing/${id}`, data),
-
-    // Tool pricing
-    getToolPricing: () =>
-      apiClient.get('/api/admin/tool-pricing'),
-    updateToolPricing: (toolName: string, data: { base_cost_usd?: number; markup_percent?: number; notes?: string; is_active?: boolean }) =>
-      apiClient.put(`/api/admin/tool-pricing/${toolName}`, data),
-    bulkMarkupToolPricing: (data: { markup_percent: number; service?: string }) =>
-      apiClient.post('/api/admin/tool-pricing/bulk-markup', data),
-    getZOStats: (params: { yearFrom: number; yearTo: number; justiceKind: string }) =>
-      apiClient.get('/api/admin/zo-stats', { params }),
-    getBulkScrapeStatus: () =>
-      apiClient.get('/api/admin/bulk-scrape-status'),
-    getInfrastructureHealth: () =>
-      apiClient.get('/api/admin/infrastructure-health'),
-
-    // Attorney profile management
-    getAttorneyProfile: (userId: string) =>
-      apiClient.get(`/api/admin/users/${userId}/attorney-profile`),
-    createAttorneyProfile: (userId: string, data: any) =>
-      apiClient.post(`/api/admin/users/${userId}/attorney-profile`, data),
-    updateAttorneyProfile: (userId: string, data: any) =>
-      apiClient.put(`/api/admin/users/${userId}/attorney-profile`, data),
-    deleteAttorneyProfile: (userId: string) =>
-      apiClient.delete(`/api/admin/users/${userId}/attorney-profile`),
-  },
-
-  // Decisions - court decision downloads from reyestr
-  decisions: {
-    checkAvailable: (docIds: string[]) =>
-      apiClient.get('/api/decisions/check-available', { params: { doc_ids: docIds.join(',') } }),
-    fetchFullText: (docId: string) =>
-      apiClient.post('/api/decisions/fetch-fulltext', { doc_id: docId }),
-    fetchFullTextBatch: (docIds: string[]) =>
-      apiClient.post('/api/decisions/fetch-fulltext-batch', { doc_ids: docIds }),
-  },
-
-  // API Keys (MCP tokens)
-  keys: {
-    list: () => apiClient.get('/api/keys'),
-    create: (data: { name: string; description?: string; expiresAt?: string }) =>
-      apiClient.post('/api/keys', data),
-    revoke: (keyId: string) => apiClient.delete(`/api/keys/${keyId}`),
-  },
-
-  // GDPR
-  gdpr: {
-    requestExport: () => apiClient.post('/api/gdpr/export'),
-    getExport: (id: string) => apiClient.get(`/api/gdpr/export/${id}`),
-    requestDeletion: (confirmation: string) =>
-      apiClient.post('/api/gdpr/delete', { confirmation }),
-    listRequests: () => apiClient.get('/api/gdpr/requests'),
-  },
-};
+export { default } from './api/client';
+export { api } from './api/index';
